Kubewarden: update deployment examples
The PSP policies got renamed to have a naming scheme that is consistent
with their GitHub repository.

This commit updates all the Kubewarden example to use the latest
versions of the PSP policies.

BTW: old files are still available over our container registry. This
change is just to make everything future proof.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>