Kubewarden: allways add hostnamespace policy
The default behavior of the PSP is to consider the `privileged`, `hostIPC`,
`hostPID` and `hostNetwork` as false. Even when the user does not set them.
Furthermore, if the value of these fields are false, `kubectl get` command
output does not show the fields. Therefore, the `psp-migration` command
is not able to migrate this fields properly.

This commit change the kubewarden script to always add the
hostnamespace policy. If the fields mentioned does not exist in the
output the default value for the Kubewarden policy is `false`.
Otherwise, the value defined in the PSP definition is used. Thus, the
PSP behavior is replicated.